+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.9.3
+- Added: Twitch Drop: Get the Cuddly Cotton Candy Grrgle Housing Decor Item June 16!
+- Added: Revelations Content Update Notes
+- Added: Twitch Drop: Get the Cuddly Cotton Candy Grrgle Housing Decor Item 16 June!
+
 ## 1.9.2
 - Added: Season of Discovery Hotfixes - 10 June
 - Added: Season of Discovery Hotfixes - June 10
@@ -390,6 +395,7 @@
 - Added external link copy and guild chat sharing for selected posts.
 - Added login toasts for recent posts.
 - Added class section shortcuts for class hotfix posts.
+